update the colormappings to handle doubles

+++ src/main/java/fact/viewer/colorMappings/ColorMapping.java
@@ -11,7 +11,7 @@ import java.awt.Color;
  */
 public interface ColorMapping {
 	
-	public Color map(float value,  float minValue,  float maxValue);
+	public Color map(double value,  double minValue,  double maxValue);
 	//public Color map(float value);
 //	public float getMinValue();
 //	public float getMaxValue();
